Extract subfilter into its own viewmodel and init it only on following
