+++ WordPress/src/main/java/org/wordpress/android/ui/reader/ReaderConstants.java
@@ -31,6 +31,7 @@ public class ReaderConstants {
 
     // intent arguments / keys
     static final String ARG_TAG = "tag";
+    static final String ARG_ORIGINAL_TAG = "original_tag";
     static final String ARG_IS_FEED = "is_feed";
     static final String ARG_BLOG_ID = "blog_id";
     static final String ARG_FEED_ID = "feed_id";
