+++ WordPress/src/test/java/org/wordpress/android/viewmodel/reader/ReaderPostListViewModelTest.kt
@@ -3,9 +3,6 @@ package org.wordpress.android.viewmodel.reader
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
-import com.nhaarman.mockitokotlin2.any
-import com.nhaarman.mockitokotlin2.times
-import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
 import kotlinx.coroutines.InternalCoroutinesApi
 import org.assertj.core.api.Assertions.assertThat
@@ -16,35 +13,19 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mock
 import org.mockito.junit.MockitoJUnitRunner
-import org.wordpress.android.TEST_DISPATCHER
-import org.wordpress.android.fluxc.model.AccountModel
 import org.wordpress.android.fluxc.model.SiteModel
 import org.wordpress.android.fluxc.store.AccountStore
 import org.wordpress.android.fluxc.store.SiteStore
 import org.wordpress.android.models.ReaderPost
-import org.wordpress.android.models.ReaderTag
-import org.wordpress.android.models.ReaderTagType.BOOKMARKED
 import org.wordpress.android.models.news.NewsItem
 import org.wordpress.android.ui.prefs.AppPrefsWrapper
-import org.wordpress.android.ui.reader.ReaderSubsActivity
 import org.wordpress.android.ui.reader.reblog.NoSite
 import org.wordpress.android.ui.reader.reblog.PostEditor
 import org.wordpress.android.ui.reader.reblog.SitePicker
 import org.wordpress.android.ui.reader.reblog.Unknown
-import org.wordpress.android.ui.reader.services.update.ReaderUpdateLogic.UpdateTask
-import org.wordpress.android.ui.reader.subfilter.ActionType.OpenLoginPage
-import org.wordpress.android.ui.reader.subfilter.ActionType.OpenSubsAtPage
-import org.wordpress.android.ui.reader.subfilter.SubfilterCategory.SITES
-import org.wordpress.android.ui.reader.subfilter.SubfilterCategory.TAGS
-import org.wordpress.android.ui.reader.subfilter.SubfilterListItem
-import org.wordpress.android.ui.reader.subfilter.SubfilterListItem.SiteAll
-import org.wordpress.android.ui.reader.subfilter.SubfilterListItem.Tag
-import org.wordpress.android.ui.reader.subfilter.SubfilterListItemMapper
 import org.wordpress.android.ui.reader.tracker.ReaderTracker
 import org.wordpress.android.ui.reader.viewmodels.ReaderPostListViewModel
 import org.wordpress.android.util.BuildConfig
-import org.wordpress.android.util.EventBusWrapper
-import java.util.EnumSet
 
 @InternalCoroutinesApi
 @RunWith(MockitoJUnitRunner::class)
@@ -54,15 +35,7 @@ class ReaderPostListViewModelTest {
     val rule = InstantTaskExecutorRule()
 
     @Mock private lateinit var observer: Observer<NewsItem>
-    @Mock private lateinit var item: NewsItem
-    /**
-     * First tag for which the card was shown.
-     */
-    @Mock private lateinit var initialTag: ReaderTag
-    @Mock private lateinit var savedTag: ReaderTag
     @Mock private lateinit var appPrefsWrapper: AppPrefsWrapper
-    @Mock private lateinit var subfilterListItemMapper: SubfilterListItemMapper
-    @Mock private lateinit var eventBusWrapper: EventBusWrapper
     @Mock private lateinit var accountStore: AccountStore
     @Mock private lateinit var readerTracker: ReaderTracker
     @Mock private lateinit var siteStore: SiteStore
@@ -72,185 +45,7 @@ class ReaderPostListViewModelTest {
 
     @Before
     fun setUp() {
-        whenever(savedTag.tagTitle).thenReturn("tag-title")
-        val tag = Tag(
-                tag = savedTag,
-                onClickAction = ::onClickActionDummy
-        )
-        val json = "{\"blogId\":0,\"feedId\":0,\"tagSlug\":\"news\",\"tagType\":1,\"type\":4}"
-
-        whenever(appPrefsWrapper.getReaderSubfilter()).thenReturn(json)
-        whenever(subfilterListItemMapper.fromJson(any(), any(), any())).thenReturn(tag)
-
-        viewModel = ReaderPostListViewModel(
-                TEST_DISPATCHER,
-                TEST_DISPATCHER,
-                appPrefsWrapper,
-                subfilterListItemMapper,
-                eventBusWrapper,
-                accountStore,
-                readerTracker,
-                siteStore
-        )
-    }
-
-    @Test
-    fun `view model change subfilter visibility as requested`() {
-        viewModel.changeSubfiltersVisibility(true)
-        assertThat(viewModel.shouldShowSubFilters.value).isEqualTo(true)
-
-        viewModel.changeSubfiltersVisibility(false)
-        assertThat(viewModel.shouldShowSubFilters.value).isEqualTo(false)
-    }
-
-    @Test
-    fun `view model returns default filter on start`() {
-        assertThat(viewModel.getCurrentSubfilterValue()).isInstanceOf(Tag::class.java)
-    }
-
-    @Test
-    fun `view model is able to set requested subfilter given a tag`() {
-        val tag = ReaderTag("", "", "", "", BOOKMARKED)
-        var item: SubfilterListItem? = null
-        viewModel.setSubfilterFromTag(tag)
-
-        viewModel.currentSubFilter.observeForever { item = it }
-
-        assertThat((item as Tag).tag).isEqualTo(tag)
-    }
-
-    @Test
-    fun `view model is able to set default subfilter`() {
-        var item: SubfilterListItem? = null
-        viewModel.setDefaultSubfilter()
-
-        viewModel.currentSubFilter.observeForever { item = it }
-
-        assertThat(item).isInstanceOf(SiteAll::class.java)
-    }
-
-    @Test
-    fun `view model updates count of matched sites and tags`() {
-        val data = hashMapOf(SITES to 3, TAGS to 25)
-        viewModel.start(initialTag, false, false)
-
-        for (testStep in data.keys) {
-            viewModel.onSubfilterPageUpdated(testStep, data.getOrDefault(testStep, 0))
-        }
-
-        assertThat(viewModel.filtersMatchCount.value).isEqualTo(data)
-    }
-
-    @Test
-    fun `when WPCOM user selects empty bottom sheet SITES cta the subs is opened on followed blogs page`() {
-        val action = OpenSubsAtPage(ReaderSubsActivity.TAB_IDX_FOLLOWED_BLOGS)
-        viewModel.onBottomSheetActionClicked(action)
-
-        assertThat(viewModel.changeBottomSheetVisibility.value!!.peekContent()).isEqualTo(false)
-        assertThat(viewModel.bottomSheetEmptyViewAction.value!!.peekContent())
-                .isEqualTo(action)
-    }
-
-    @Test
-    fun `when WPCOM user selects empty bottom sheet TAGS cta the subs is opened on followed tags page`() {
-        val action = OpenSubsAtPage(ReaderSubsActivity.TAB_IDX_FOLLOWED_TAGS)
-        viewModel.onBottomSheetActionClicked(action)
-
-        assertThat(viewModel.changeBottomSheetVisibility.value!!.peekContent()).isEqualTo(false)
-        assertThat(viewModel.bottomSheetEmptyViewAction.value!!.peekContent())
-                .isEqualTo(action)
-    }
-
-    @Test
-    fun `when self-hosted user selects empty bottom sheet cta the me page is opened`() {
-        val action = OpenLoginPage
-        viewModel.onBottomSheetActionClicked(action)
-
-        assertThat(viewModel.changeBottomSheetVisibility.value!!.peekContent()).isEqualTo(false)
-        assertThat(viewModel.bottomSheetEmptyViewAction.value!!.peekContent())
-                .isEqualTo(action)
-    }
-
-    @Test
-    fun `when user id changed a tags and blogs update is triggered and default subfilter is set`() {
-        whenever(appPrefsWrapper.getLastReaderKnownUserId()).thenReturn(0)
-        whenever(appPrefsWrapper.getLastReaderKnownAccessTokenStatus()).thenReturn(true)
-        whenever(accountStore.hasAccessToken()).thenReturn(true)
-        val account = AccountModel()
-        account.userId = 100
-        whenever(accountStore.account).thenReturn(account)
-
-        viewModel.onUserComesToReader()
-
-        verify(appPrefsWrapper, times(1)).setLastReaderKnownUserId(any())
-        verify(appPrefsWrapper, times(0)).setLastReaderKnownAccessTokenStatus(any())
-
-        assertThat(viewModel.updateTagsAndSites.value!!.peekContent()).isEqualTo(
-                EnumSet.of(
-                        UpdateTask.TAGS,
-                        UpdateTask.FOLLOWED_BLOGS
-                )
-        )
-
-        assertThat(viewModel.currentSubFilter.value).isInstanceOf(SiteAll::class.java)
-    }
-
-    @Test
-    fun `when user switches from wpcom and self-hosted an update is triggered and default subfilter is set`() {
-        whenever(appPrefsWrapper.getLastReaderKnownUserId()).thenReturn(100)
-        whenever(appPrefsWrapper.getLastReaderKnownAccessTokenStatus()).thenReturn(true)
-        whenever(accountStore.hasAccessToken()).thenReturn(false)
-
-        viewModel.onUserComesToReader()
-
-        verify(appPrefsWrapper, times(0)).setLastReaderKnownUserId(any())
-        verify(appPrefsWrapper, times(1)).setLastReaderKnownAccessTokenStatus(any())
-
-        assertThat(viewModel.updateTagsAndSites.value!!.peekContent()).isEqualTo(
-                EnumSet.of(
-                        UpdateTask.TAGS,
-                        UpdateTask.FOLLOWED_BLOGS
-                )
-        )
-
-        assertThat(viewModel.currentSubFilter.value).isInstanceOf(SiteAll::class.java)
-    }
-
-    @Test
-    fun `when user id do not change nothing happens`() {
-        whenever(appPrefsWrapper.getLastReaderKnownUserId()).thenReturn(100)
-        whenever(appPrefsWrapper.getLastReaderKnownAccessTokenStatus()).thenReturn(true)
-        whenever(accountStore.hasAccessToken()).thenReturn(true)
-        val account = AccountModel()
-        account.userId = 100
-        whenever(accountStore.account).thenReturn(account)
-
-        viewModel.onUserComesToReader()
-
-        verify(appPrefsWrapper, times(0)).setLastReaderKnownUserId(any())
-        verify(appPrefsWrapper, times(0)).setLastReaderKnownAccessTokenStatus(any())
-
-        assertThat(viewModel.updateTagsAndSites.value).isEqualTo(null)
-
-        // we didn't call start so noone should have changed the value
-        assertThat(viewModel.currentSubFilter.value).isEqualTo(null)
-    }
-
-    @Test
-    fun `when user remains self-hosted nothing happens`() {
-        whenever(appPrefsWrapper.getLastReaderKnownUserId()).thenReturn(100)
-        whenever(appPrefsWrapper.getLastReaderKnownAccessTokenStatus()).thenReturn(false)
-        whenever(accountStore.hasAccessToken()).thenReturn(false)
-
-        viewModel.onUserComesToReader()
-
-        verify(appPrefsWrapper, times(0)).setLastReaderKnownUserId(any())
-        verify(appPrefsWrapper, times(0)).setLastReaderKnownAccessTokenStatus(any())
-
-        assertThat(viewModel.updateTagsAndSites.value).isEqualTo(null)
-
-        // we didn't call start so noone should have changed the value
-        assertThat(viewModel.currentSubFilter.value).isEqualTo(null)
+        viewModel = ReaderPostListViewModel(readerTracker, siteStore)
     }
 
     @Test
@@ -359,9 +154,5 @@ class ReaderPostListViewModelTest {
             val state = viewModel.reblogState.value?.peekContent()
             assertThat(state).isInstanceOf(Unknown::class.java)
         }
-    }
-
-    private fun onClickActionDummy(filter: SubfilterListItem) {
-        return
     }
 }
